Add placement of TOP slabs below existing TOP slabs
Moves a top slab that was placed into an existing top slab below said
existing top slab. This is weird though as sometimes it just does not
work. As in, you place a top slab, then place another top slab into the
existing slab, and it chooses not to work. Seconds later, after moving
around and/or replacing said initial slab, it will work again. I have
no idea why this is. I will #research this.